Fixed time initialization to fix initial TPF bug.
With the old mechanism, the first TPF value would be reported as a number of seconds (often ~10), since the last start/end times would be set to zero and the current elapsed millisecond clock (elapsed time since JVM start) was used. This in effect gave an initial TPF relative to the start of the JVM, not zero as the first TPF should be.